Rework transaction management.
This commit changes the way that transactions are handled in the db.
Before this commit, all the transactions were persisted in a single
file, which eventually would become a problem due to the amount of
transactions that can be generated with the passage of time. Thus,
the mechanism to manage the transactions has been changed. Now, the
transactions are persisted in files according to the date of the
transaction. That is, the year of the transaction defines the
directory in which the file would be stored and the month defines
the file in which the transaction will be persisted. An example path
to a file would be /home/user/.fintool/2020/01.csv. These files are
now called collections, since they represent a collection of
transactions.

Signed-off-by: Isaias Ramirez <soote1.ir@gmail.com>